+++ api/python/arrus/session.py
@@ -94,15 +94,24 @@ class Session(AbstractSession):
         medium = self._context.medium
         seq = scheme.tx_rx_sequence
         processing = scheme.processing
+        constants = scheme.constants
 
         kernel_context = self._create_kernel_context(
             seq,
             us_device_dto,
             medium,
-            scheme.digital_down_conversion)
-        raw_seq = arrus.kernels.get_kernel(type(seq))(kernel_context)
+            scheme.digital_down_conversion,
+            constants
+        )
+        conversion_results = arrus.kernels.get_kernel(type(seq))(kernel_context)
+        raw_seq = conversion_results.sequence
+        tx_delay_constants = conversion_results.constants
 
-        actual_scheme = dataclasses.replace(scheme, tx_rx_sequence=raw_seq)
+        actual_scheme = dataclasses.replace(
+            scheme,
+            tx_rx_sequence=raw_seq,
+            constants=tx_delay_constants
+        )
         core_scheme = arrus.utils.core.convert_to_core_scheme(actual_scheme)
         upload_result = self._session_handle.upload(core_scheme)
 
@@ -115,7 +124,7 @@ class Session(AbstractSession):
         # -- Constant metadata
         # --- Frame acquisition context
         fac = self._create_frame_acquisition_context(
-            seq, raw_seq, us_device_dto, medium)
+            seq, raw_seq, us_device_dto, medium, tx_delay_constants)
 
         buffer = arrus.framework.DataBuffer(buffer_handle)
         input_shape = buffer.elements[0].data.shape
@@ -291,15 +300,20 @@ class Session(AbstractSession):
             devices[(arrus.core.DeviceType_GPU, 0)] = arrus.devices.gpu.GPU(0)
         return devices
 
-    def _create_kernel_context(self, seq, device, medium, hardware_ddc):
+    def _create_kernel_context(self, seq, device, medium, hardware_ddc,
+                               constants):
         return arrus.kernels.kernel.KernelExecutionContext(
             device=device, medium=medium, op=seq, custom={},
-            hardware_ddc=hardware_ddc
+            hardware_ddc=hardware_ddc,
+            constants=constants
         )
 
-    def _create_frame_acquisition_context(self, seq, raw_seq, device, medium):
+    def _create_frame_acquisition_context(self, seq, raw_seq, device, medium,
+                                          constants):
         return arrus.metadata.FrameAcquisitionContext(
             device=device, sequence=seq, raw_sequence=raw_seq,
-            medium=medium, custom_data={})
+            medium=medium, custom_data={},
+            constants=constants
+        )