+++ arrus/core/devices/us4r/Us4RFactoryImpl.h
@@ -18,6 +18,7 @@
 
 #include "arrus/core/devices/us4r/external/ius4oem/IUs4OEMFactory.h"
 #include "arrus/core/devices/us4r/hv/HighVoltageSupplierFactory.h"
+#include "arrus/core/devices/us4r/backplane/DigitalBackplaneFactory.h"
 #include "arrus/core/devices/us4r/Us4RSettingsConverter.h"
 
 namespace arrus::devices {
@@ -30,14 +31,16 @@ class Us4RFactoryImpl : public Us4RFactory {
                     std::unique_ptr<IUs4OEMFactory> ius4oemFactory,
                     std::unique_ptr<IUs4OEMInitializer> ius4oemInitializer,
                     std::unique_ptr<Us4RSettingsConverter> us4RSettingsConverter,
-                    std::unique_ptr<HighVoltageSupplierFactory> hvFactory)
+                    std::unique_ptr<HighVoltageSupplierFactory> hvFactory,
+                    std::unique_ptr<DigitalBackplaneFactory> backplaneFactory)
         : ius4oemFactory(std::move(ius4oemFactory)),
           ius4oemInitializer(std::move(ius4oemInitializer)),
           us4oemFactory(std::move(us4oemFactory)),
           us4RSettingsConverter(std::move(us4RSettingsConverter)),
           probeAdapterFactory(std::move(adapterFactory)),
           probeFactory(std::move(probeFactory)),
-          hvFactory(std::move(hvFactory)) {}
+          hvFactory(std::move(hvFactory)),
+          backplaneFactory(std::move(backplaneFactory)){}
 
     Us4R::Handle getUs4R(Ordinal ordinal, const Us4RSettings &settings) override {
         DeviceId id(DeviceType::Us4R, ordinal);
@@ -74,7 +77,8 @@ class Us4RFactoryImpl : public Us4RFactory {
             // verify if the generated us4oemSettings.channelsMask is equal to us4oemChannelsMask field
             validateChannelsMasks(us4OEMSettings, settings.getUs4OEMChannelsMask());
 
-            auto[us4oems, masterIUs4OEM] = getUs4OEMs(us4OEMSettings, settings.isExternalTrigger());
+            auto[us4oems, masterIUs4OEM] =
+                getUs4OEMs(us4OEMSettings, settings.isExternalTrigger(), probeAdapterSettings.getIOSettings());
             std::vector<Us4OEMImplBase::RawHandle> us4oemPtrs(us4oems.size());
             std::transform(std::begin(us4oems), std::end(us4oems), std::begin(us4oemPtrs),
                 [](const Us4OEMImplBase::Handle &ptr) { return ptr.get(); });
@@ -88,19 +92,22 @@ class Us4RFactoryImpl : public Us4RFactory {
                 ius4oems.push_back(us4oem->getIUs4oem());
             }
 
-            auto hv = getHV(settings.getHVSettings(), ius4oems);
+            auto backplane = getBackplane(settings.getDigitalBackplaneSettings(), settings.getHVSettings(), ius4oems);
+            auto hv = getHV(settings.getHVSettings(), ius4oems, backplane);
             return std::make_unique<Us4RImpl>(id, std::move(us4oems), adapter, probe, std::move(hv), rxSettings,
-                                              settings.getChannelsMask());
+                                              settings.getChannelsMask(), std::move(backplane));
         } else {
             // Custom Us4OEMs only
-            auto[us4oems, masterIUs4OEM] = getUs4OEMs(settings.getUs4OEMSettings(), false);
+            auto[us4oems, masterIUs4OEM] = getUs4OEMs(settings.getUs4OEMSettings(), false, us4r::IOSettings());
             std::vector<IUs4OEM*> ius4oems;
             for(auto &us4oem: us4oems) {
                 ius4oems.push_back(us4oem->getIUs4oem());
             }
 
-            auto hv = getHV(settings.getHVSettings(), ius4oems);
-            return std::make_unique<Us4RImpl>(id, std::move(us4oems), std::move(hv), settings.getChannelsMask());
+            auto backplane = getBackplane(settings.getDigitalBackplaneSettings(), settings.getHVSettings(), ius4oems);
+            auto hv = getHV(settings.getHVSettings(), ius4oems, backplane);
+            return std::make_unique<Us4RImpl>(id, std::move(us4oems), std::move(hv), settings.getChannelsMask(),
+                                              std::move(backplane));
         }
     }
 
@@ -142,7 +149,7 @@ class Us4RFactoryImpl : public Us4RFactory {
      * @return a pair: us4oems, master ius4oem
      */
     std::pair<std::vector<Us4OEMImplBase::Handle>, IUs4OEM *>
-    getUs4OEMs(const std::vector<Us4OEMSettings> &us4oemCfgs, bool isExternalTrigger) {
+    getUs4OEMs(const std::vector<Us4OEMSettings> &us4oemCfgs, bool isExternalTrigger, const us4r::IOSettings& io) {
         ARRUS_REQUIRES_AT_LEAST(us4oemCfgs.size(), 1,"At least one us4oem should be configured.");
         auto nUs4oems = static_cast<Ordinal>(us4oemCfgs.size());
 
@@ -169,21 +176,68 @@ class Us4RFactoryImpl : public Us4RFactory {
         ARRUS_REQUIRES_EQUAL(ius4oems.size(), us4oemCfgs.size(),
                              ArrusException("Values are not equal: ius4oem size, us4oem settings size"));
 
+        // Determine which OEMs must acquire RX nops, for pulse counter capability.
+        std::unordered_set<Ordinal> pulseCounterOems;
+        if(io.hasFrameMetadataCapability())  {
+            pulseCounterOems = io.getFrameMetadataCapabilityOEMs();
+        }
+        else {
+            // By default us4OEM:0 is the pulse counter.
+            pulseCounterOems.insert(Ordinal(0));
+        }
         for (unsigned i = 0; i < ius4oems.size(); ++i) {
             // TODO(Us4R-10) use ius4oem->GetDeviceID() as an ordinal number, instead of value of i
-            us4oems.push_back(us4oemFactory->getUs4OEM(static_cast<ChannelIdx>(i), ius4oems[i], us4oemCfgs[i],
-                                                       isExternalTrigger));
+            auto ordinal = static_cast<Ordinal>(i);
+            us4oems.push_back(us4oemFactory->getUs4OEM(
+                static_cast<Ordinal>(i),
+                ius4oems[i],
+                us4oemCfgs[i],
+                isExternalTrigger,
+                setContains(pulseCounterOems, ordinal) // accept RX nops?
+                // NOTE: the above should be consistent with the ProbeAdapterImpl::frameMetadataOem
+            ));
         }
+        initCapabilities(us4oems, io);
         return {std::move(us4oems), master};
     }
 
-    std::vector<HighVoltageSupplier::Handle> getHV(const std::optional<HVSettings> &settings, std::vector<IUs4OEM *> &us4oems) {
+    std::vector<HighVoltageSupplier::Handle> getHV(const std::optional<HVSettings> &settings,
+                                                   std::vector<IUs4OEM *> &us4oems,
+                                                   const std::optional<DigitalBackplane::Handle> &backplane) {
         if (settings.has_value()) {
             const auto &hvSettings = settings.value();
-            return hvFactory->getHighVoltageSupplier(hvSettings, us4oems);
+            return hvFactory->getHighVoltageSupplier(hvSettings, us4oems, backplane);
         } else {
             std::vector<HighVoltageSupplier::Handle> empty;
             return empty;
+        }
+    }
+
+    std::optional<DigitalBackplane::Handle> getBackplane(
+        const std::optional<DigitalBackplaneSettings> &dbarSettings,
+        const std::optional<HVSettings> &hvSettings,
+        std::vector<IUs4OEM *> &us4oems) {
+
+        if(dbarSettings.has_value()) {
+            return backplaneFactory->getDigitalBackplane(dbarSettings.value(), us4oems);
+        }
+        else if (hvSettings.has_value()) {
+            // Fallback option: try to determine HV model based on the HV in use.
+            return backplaneFactory->getDigitalBackplane(hvSettings.value(), us4oems);
+        } else {
+            return std::nullopt;
+        }
+    }
+
+    void initCapabilities(const Us4RImpl::Us4OEMs& us4oems, const us4r::IOSettings settings) {
+        if(settings.hasProbeConnectedCheckCapability()) {
+            auto addr = settings.getProbeConnectedCheckCapabilityAddress();
+            if(addr.getUs4OEM() == 0){
+                us4oems.at(addr.getUs4OEM())->getIUs4oem()->EnableProbeCheck(addr.getIO());
+            }
+            else {
+                throw arrus::IllegalArgumentException("Probe check functionality must be connected to us4OEM #0");
+            }
         }
     }
 
@@ -194,6 +248,7 @@ class Us4RFactoryImpl : public Us4RFactory {
     std::unique_ptr<ProbeAdapterFactory> probeAdapterFactory;
     std::unique_ptr<ProbeFactory> probeFactory;
     std::unique_ptr<HighVoltageSupplierFactory> hvFactory;
+    std::unique_ptr<DigitalBackplaneFactory> backplaneFactory;
 };
 
 }