+++ arrus/core/devices/us4r/Us4RImpl.cpp
@@ -6,6 +6,7 @@
 #include <chrono>
 #include <memory>
 #include <thread>
+#include <future>
 
 #define ARRUS_ASSERT_RX_SETTINGS_SET()                                                                                 \
     if (!rxSettings.has_value()) {                                                                                     \
@@ -38,18 +39,26 @@ UltrasoundDevice *Us4RImpl::getDefaultComponent() {
 }
 
 Us4RImpl::Us4RImpl(const DeviceId &id, Us4OEMs us4oems, std::vector<HighVoltageSupplier::Handle> hv,
-                   std::vector<unsigned short> channelsMask)
-    : Us4R(id), logger{getLoggerFactory()->getLogger()}, us4oems(std::move(us4oems)), hv(std::move(hv)),
-      channelsMask(std::move(channelsMask)) {
+                   std::vector<unsigned short> channelsMask, std::optional<DigitalBackplane::Handle> backplane)
+    : Us4R(id), logger{getLoggerFactory()->getLogger()}, us4oems(std::move(us4oems)),
+      digitalBackplane(std::move(backplane)),
+      hv(std::move(hv)),
+      channelsMask(std::move(channelsMask))
+{
     INIT_ARRUS_DEVICE_LOGGER(logger, id.toString());
 }
 
 Us4RImpl::Us4RImpl(const DeviceId &id, Us4RImpl::Us4OEMs us4oems, ProbeAdapterImplBase::Handle &probeAdapter,
                    ProbeImplBase::Handle &probe, std::vector<HighVoltageSupplier::Handle> hv,
-                   const RxSettings &rxSettings, std::vector<unsigned short> channelsMask)
+                   const RxSettings &rxSettings, std::vector<unsigned short> channelsMask,
+                   std::optional<DigitalBackplane::Handle> backplane)
     : Us4R(id), logger{getLoggerFactory()->getLogger()}, us4oems(std::move(us4oems)),
-      probeAdapter(std::move(probeAdapter)), probe(std::move(probe)), hv(std::move(hv)), rxSettings(rxSettings),
-      channelsMask(std::move(channelsMask)) {
+      probeAdapter(std::move(probeAdapter)), probe(std::move(probe)),
+      digitalBackplane(std::move(backplane)),
+      hv(std::move(hv)),
+      rxSettings(rxSettings),
+      channelsMask(std::move(channelsMask))
+{
     INIT_ARRUS_DEVICE_LOGGER(logger, id.toString());
 }
 
@@ -87,7 +96,7 @@ std::vector<std::pair <std::string,float>> Us4RImpl::logVoltages(bool isHV256) {
         //Currently OEM+ does not support internal voltage measurement - skip
     }
 
-    
+
 
     return voltages;
 }
@@ -140,10 +149,31 @@ void Us4RImpl::setVoltage(Voltage voltage) {
             ::arrus::format("Unaccepted voltage '{}', should be in range: [{}, {}]", voltage, minVoltage, maxVoltage));
     }
 
+    bool isHVPS = true;
+
     for(uint8_t n = 0; n < hv.size(); n++) {
-        hv[n]->setVoltage(voltage);
+        auto &hvModel = this->hv[n]->getModelId();
+        if(hvModel.getName() != "us4oemhvps") {
+            isHVPS = false;
+            break;
+        }
     }
-    
+
+    if(isHVPS) {
+        std::vector<std::future<void>> futures;
+        for (uint8_t n = 0; n < hv.size(); n++) {
+            futures.push_back(std::async(std::launch::async, &HighVoltageSupplier::setVoltage, hv[n].get(), voltage));
+        }
+        for (auto &future : futures) {
+            future.wait();
+        }
+    }
+    else {
+        for(uint8_t n = 0; n < hv.size(); n++) {
+            hv[n]->setVoltage(voltage);
+        }
+    }
+
 
     //Wait to stabilise voltage output
     std::this_thread::sleep_for(std::chrono::milliseconds(1000));
@@ -215,7 +245,6 @@ std::pair<Buffer::SharedHandle, arrus::session::Metadata::SharedHandle>
 Us4RImpl::upload(const ::arrus::ops::us4r::Scheme &scheme) {
     auto &outputBufferSpec = scheme.getOutputBuffer();
     auto rxBufferNElements = scheme.getRxBufferSize();
-    auto &seq = scheme.getTxRxSequence();
     auto workMode = scheme.getWorkMode();
 
     unsigned hostBufferNElements = outputBufferSpec.getNumberOfElements();
@@ -235,8 +264,11 @@ Us4RImpl::upload(const ::arrus::ops::us4r::Scheme &scheme) {
     }
     // Upload and register buffers.
     bool useTriggerSync = workMode == Scheme::WorkMode::HOST || workMode == Scheme::WorkMode::MANUAL;
+
+    auto &seq = scheme.getTxRxSequence();
+
     auto [rxBuffer, fcm] = uploadSequence(seq, rxBufferNElements, seq.getNRepeats(), useTriggerSync,
-                                          scheme.getDigitalDownConversion());
+                                          scheme.getDigitalDownConversion(), scheme.getConstants());
     ARRUS_REQUIRES_TRUE(!rxBuffer->empty(), "Us4R Rx buffer cannot be empty.");
 
     // Calculate how much of the data each Us4OEM produces.
@@ -341,7 +373,8 @@ Us4RImpl::~Us4RImpl() {
 
 std::tuple<Us4RBuffer::Handle, FrameChannelMapping::Handle>
 Us4RImpl::uploadSequence(const TxRxSequence &seq, uint16 bufferSize, uint16 batchSize, bool triggerSync,
-                         const std::optional<ops::us4r::DigitalDownConversion> &ddc) {
+                         const std::optional<ops::us4r::DigitalDownConversion> &ddc,
+                         const std::vector<framework::NdArray> &txDelayProfiles) {
     std::vector<TxRxParameters> actualSeq;
     // Convert to intermediate representation (TxRxParameters).
     size_t opIdx = 0;
@@ -357,7 +390,7 @@ Us4RImpl::uploadSequence(const TxRxSequence &seq, uint16 bufferSize, uint16 batc
         ++opIdx;
     }
     return getProbeImpl()->setTxRxSequence(actualSeq, seq.getTgcCurve(), bufferSize, batchSize, seq.getSri(),
-                                           triggerSync, ddc);
+                                           triggerSync, ddc, txDelayProfiles);
 }
 
 void Us4RImpl::trigger() { this->getDefaultComponent()->syncTrigger(); }
@@ -760,6 +793,44 @@ std::function<void()> Us4RImpl::createOnTransferOverflowCallback(
         };
     default:
         throw ::arrus::IllegalArgumentException("Unsupported work mode.");
+    }
+}
+
+const char *Us4RImpl::getBackplaneSerialNumber() {
+    if(!this->digitalBackplane.has_value()) {
+        throw arrus::IllegalArgumentException("No backplane defined.");
+    }
+    return this->digitalBackplane->get()->getSerialNumber();
+}
+
+const char *Us4RImpl::getBackplaneRevision() {
+    if(!this->digitalBackplane.has_value()) {
+        throw arrus::IllegalArgumentException("No backplane defined.");
+    }
+    return this->digitalBackplane->get()->getRevisionNumber();
+}
+
+void Us4RImpl::setParameters(const Parameters &params) {
+    for(auto &item: params.items()) {
+        auto &key = item.first;
+        auto value = item.second;
+        logger->log(LogSeverity::INFO, format("Setting value {} to {}", value, key));
+        if(key != "/sequence:0/txFocus") {
+            throw ::arrus::IllegalArgumentException("Currently Us4R supports only sequence:0/txFocus parameter.");
+        }
+        this->us4oems[0]->getIUs4oem()->TriggerStop();
+        try {
+            for(auto &us4oem: us4oems) {
+                us4oem->getIUs4oem()->SetTxDelays(value);
+            }
+	} 
+	catch(...) {
+            // Try resume.
+            this->us4oems[0]->getIUs4oem()->TriggerStart();
+	    throw;
+	}
+	// Everything OK, resume.
+        this->us4oems[0]->getIUs4oem()->TriggerStart();
     }
 }