+++ arrus/core/devices/us4r/Us4RSettingsConverterImpl.h
@@ -132,7 +132,9 @@ public:
                 ProbeAdapterSettings(
                     probeAdapterSettings.getModelId(),
                     probeAdapterSettings.getNumberOfChannels(),
-                    adapterChannelMapping)};
+                    adapterChannelMapping,
+                    probeAdapterSettings.getIOSettings()
+                )};
     }
 
 private: