+++ arrus/core/devices/us4r/us4oem/Us4OEMImplTest.cpp
@@ -77,6 +77,7 @@ protected:
             channelMapping, rxSettings,
             std::unordered_set<uint8>(),
             Us4OEMSettings::ReprogrammingMode::SEQUENTIAL,
+            false,
             false
         );
     }
@@ -364,6 +365,7 @@ protected:
             channelMapping, rxSettings,
             std::unordered_set<uint8>(),
             Us4OEMSettings::ReprogrammingMode::SEQUENTIAL,
+            false,
             false
         );
     }
@@ -432,7 +434,7 @@ TEST_F(Us4OEMImplEsaote3LikeTest, SetsCorrectRxTimeAndDelay1) {
             (x.sampleRange = sampleRange))
             .getTxRxParameters()
     };
-    EXPECT_CALL(*ius4oemPtr, SetRxDelay(Us4OEMImpl::RX_DELAY, 0));
+    EXPECT_CALL(*ius4oemPtr, SetRxDelay(0.0f, 0)); // Note: the value 0.0 is the default value
     uint32 nSamples = sampleRange.end() - sampleRange.start();
     float minimumRxTime = float(nSamples) / Us4OEMImpl::SAMPLING_FREQUENCY;
     EXPECT_CALL(*ius4oemPtr, SetRxTime(Ge(minimumRxTime), 0));
@@ -452,7 +454,7 @@ TEST_F(Us4OEMImplEsaote3LikeTest, SetsCorrectRxTimeAndDelay2) {
             (x.sampleRange = sampleRange))
             .getTxRxParameters()
     };
-    EXPECT_CALL(*ius4oemPtr, SetRxDelay(Us4OEMImpl::RX_DELAY, 0));
+    EXPECT_CALL(*ius4oemPtr, SetRxDelay(0.0f, 0)); // Note: the default value of TxRxParameters
     uint32 nSamples = sampleRange.end() - sampleRange.start();
     float minimumRxTime = float(nSamples) / Us4OEMImpl::SAMPLING_FREQUENCY;
     EXPECT_CALL(*ius4oemPtr, SetRxTime(Ge(minimumRxTime), 0));
@@ -706,7 +708,9 @@ protected:
             std::move(ius4oem), activeChannelGroups,
             channelMapping, rxSettings, channelsMask,
             Us4OEMSettings::ReprogrammingMode::SEQUENTIAL,
-            false);
+            false,
+            false
+            );
 
     }
 
@@ -751,7 +755,7 @@ TEST_F(Us4OEMImplEsaote3ChannelsMaskTest, DoesNothingWithAperturesWhenNoChannelM
     EXPECT_CALL(*ius4oemPtr, SetRxAperture(expectedRxAperture, 0));
     EXPECT_CALL(*ius4oemPtr, SetTxAperture(expectedTxAperture, 0));
     for(int i = 0; i < expectedTxDelays.size(); ++i) {
-        EXPECT_CALL(*ius4oemPtr, SetTxDelay(i, expectedTxDelays[i], 0));
+        EXPECT_CALL(*ius4oemPtr, SetTxDelay(i, expectedTxDelays[i], 0, 0));
     }
 
     SET_TX_RX_SEQUENCE(us4oem, seq);
@@ -791,7 +795,7 @@ TEST_F(Us4OEMImplEsaote3ChannelsMaskTest, MasksProperlyASingleChannel) {
     EXPECT_CALL(*ius4oemPtr, SetRxAperture(expectedRxAperture, 0));
     EXPECT_CALL(*ius4oemPtr, SetTxAperture(expectedTxAperture, 0));
     for(int i = 0; i < expectedTxDelays.size(); ++i) {
-        EXPECT_CALL(*ius4oemPtr, SetTxDelay(i, expectedTxDelays[i], 0));
+        EXPECT_CALL(*ius4oemPtr, SetTxDelay(i, expectedTxDelays[i], 0, 0));
     }
 
     auto [buffer, fcm] = SET_TX_RX_SEQUENCE(us4oem, seq);
@@ -990,6 +994,7 @@ protected:
                 channelMapping, rxSettings,
                 std::unordered_set<uint8>({}),
                 reprogrammingMode,
+                false,
                 false
         );