+++ docs/content/misc/release_notes.rst
@@ -10,6 +10,9 @@ Release notes
 
     - us4R: exposed SYNC mode
     - Implemented File device (for simulated mode) #221, #ARRUS-248.
+    - us4R: exposed us4OEM and digital backplane serial number and revision #ARRUS-267.
+    - Added Session::getCurrentState method, which returns the current state of the session (e.g if the devices are running or not).
+    - us4R: Provided the possibility to explicitly specify what digital backplane is used #US4R-352.
     
 - Python API:
 
@@ -20,6 +23,8 @@ Release notes
     - Created cupy -> DL pack operator.
     - Created an example for file device (simulated mode) #221, #ARRUS-248.
     - Added spacing to echo data description #295 #ARRUS-251.
+    - us4R: exposed us4OEM and digital backplane serial number and revision #ARRUS-267.
+    - us4R: provided the possibility to dynamically change TX focus, #US4R-326.
 
 
 0.8.x